confirming to cloud model

+++ src/main/java/com/aws/iot/evergreen/deployment/model/ComponentUpdatePolicy.java
@@ -24,7 +24,7 @@ import lombok.ToString;
 @ToString
 public class ComponentUpdatePolicy {
     @JsonProperty("Timeout")
-    private long timeout;
+    private Integer timeout;
     @JsonProperty("ComponentUpdatePolicyAction")
     private ComponentUpdatePolicyAction componentUpdatePolicyAction;
 }
